タイトルセクションコンポーネント追加
